chore: quarantine orphan + non-orth drivers to legacy/ (Phase D, L1)
Move 18 unbuilt-needed drivers out of the build graph without deleting them:
fifteen orphaned/unported single-purpose drivers plus the three standalone
non-orthogonal drivers (the Santiago comparison corpus). None is referenced by any
of the 24 ctests.

Mechanics per file: `git mv src/<f>.cpp legacy/<f>.cpp` (history preserved); its
`LINQT_APPS` line in src/CMakeLists.txt is commented out (`# [LEGACY] ...`), leaving
the old target->source mapping on record; a two-line banner is prepended marking it
NOT BUILT / reference-only. `legacy/` has no add_subdirectory and is never linked;
legacy/README.md states the read-only policy.

Santiago gate: the three non-orth drivers were diffed against merge-base 5cc0924
(one identical, one changed-history-retained, two newer) -- archived, not deleted;
the core non-orth logic stays in the library as an intentional throwing placeholder
with the working impl preserved at 5cc0924. L2 oracle drivers still wired into ctest
stay in src/ (deferred).

Baseline HEAD 5116491; reconfigure + build of remaining targets OK, ctest green
(24/24), golden hashes unchanged.
