+++ legacy/localDensitiesFromChebmom.cpp
@@ -1,3 +1,5 @@
+// [LEGACY -- NOT BUILT] previous-working localDensitiesFromChebmom. Reference only.
+// Supported path: lsquant <cmd>  |  port + golden-gate before any reuse.
 
 // C & C++ libraries
 #include <iostream>		/* for std::cout mostly */
