+++ src/CMakeLists.txt
@@ -55,35 +55,35 @@ set(LINQT_APPS
     "inline_compute-kpm-nonEqOp|inline_compute-kpm-nonEqOp.cpp"
     "inline_compute-kpm-CorrOp|inline_compute-kpm-CorrOp.cpp"
     "inline_compute-kpm-spectralOp|inline_compute-kpm-spectralOp.cpp"
-    "inline_kpm_S-inversion|inline_kpm-S-inversion.cpp"
-    "inline_compute-kpm-GetDeltaPhi|inline_compute-kpm-GetDeltaPhi.cpp"
-    "inline_compute-kpm-LocalDOS|inline_compute-kpm-LocalDOS.cpp"
-    "inline_compute-kpm-spacialresolution|inline_compute-kpm-spacialresolution.cpp"
+    # [LEGACY] "inline_kpm_S-inversion|inline_kpm-S-inversion.cpp"
+    # [LEGACY] "inline_compute-kpm-GetDeltaPhi|inline_compute-kpm-GetDeltaPhi.cpp"
+    # [LEGACY] "inline_compute-kpm-LocalDOS|inline_compute-kpm-LocalDOS.cpp"
+    # [LEGACY] "inline_compute-kpm-spacialresolution|inline_compute-kpm-spacialresolution.cpp"
     "inline_spectralFunctionFromChebmom|spectralFunctionFromChebmom.cpp"
-    "inline_localDensitiesFromChebmom|localDensitiesFromChebmom.cpp"
+    # [LEGACY] "inline_localDensitiesFromChebmom|localDensitiesFromChebmom.cpp"
     "inline_kuboGreenwoodFromChebmom|kuboGreenwoodFromChebmom.cpp"
-    "inline_kuboBastinIFromChebmom|kuboBastinIFromChebmom.cpp"
-    "inline_kuboBastinIIFromChebmom|kuboBastinIIFromChebmom.cpp"
+    # [LEGACY] "inline_kuboBastinIFromChebmom|kuboBastinIFromChebmom.cpp"
+    # [LEGACY] "inline_kuboBastinIIFromChebmom|kuboBastinIIFromChebmom.cpp"
     "inline_kuboBastinFromChebmom|kuboBastinFromChebmom.cpp"
     "inline_kuboBastinSeaFromChebmom|kuboBastinSeaFromChebmom.cpp"
     "inline_kuboBastinSurfFromChebmom|kuboBastinSurfFromChebmom.cpp"
     "inline_timeCorrelationsFromChebmom|timeCorrelationsFromChebmom.cpp"
-    "inline_compute-Kubo-kpm-FFT|inline_compute-Kubo-kpm-FFT.cpp"
-    "inline_opticalConductivityFromChebmom|opticalConductivityFromChebmom.cpp"
-    "inline_spectralFunctionFromChebmom_FFTgrid|spectralFunctionFromChebmom_FFTgrid.cpp"
-    "inline_compute-kpm-TimeEvProjetedOpWF|inline_compute-kpm-TimeEvProjetedOp-Projecting_WF.cpp"
-    "inline_spectralFunctionFromChebmom_Lorentz|spectralFunctionFromChebmom_Lorentz.cpp"
+    # [LEGACY] "inline_compute-Kubo-kpm-FFT|inline_compute-Kubo-kpm-FFT.cpp"
+    # [LEGACY] "inline_opticalConductivityFromChebmom|opticalConductivityFromChebmom.cpp"
+    # [LEGACY] "inline_spectralFunctionFromChebmom_FFTgrid|spectralFunctionFromChebmom_FFTgrid.cpp"
+    # [LEGACY] "inline_compute-kpm-TimeEvProjetedOpWF|inline_compute-kpm-TimeEvProjetedOp-Projecting_WF.cpp"
+    # [LEGACY] "inline_spectralFunctionFromChebmom_Lorentz|spectralFunctionFromChebmom_Lorentz.cpp"
     "inline_kuboGreenwoodFromChebmom_FFTgrid|kuboGreenwoodFromChebmom_FFTgrid.cpp"
     "inline_compute-kpm-MeanSquareDisplacement|inline_compute-kpm-MeanSquareDisplacement.cpp"
     "inline_kuboBastinFromChebmom_FFTgrid|kuboBastinFromChebmom_FFTgrid.cpp"
     "inline_compute-kpm-TimeEvProjetedOp|inline_compute-kpm-TimeEvProjetedOp.cpp"
-    "inline_compute-kpm-spectralOp-nonOrth|inline_compute-kpm-spectralOp-nonOrth.cpp"
-    "inline_compute-kpm-TimeEvOp|inline_compute-kpm-TimeEvOp.cpp"
-    "kpm-nonOrth-precision-test|kpm-nonOrth-precision-test.cpp"
+    # [LEGACY] "inline_compute-kpm-spectralOp-nonOrth|inline_compute-kpm-spectralOp-nonOrth.cpp"
+    # [LEGACY] "inline_compute-kpm-TimeEvOp|inline_compute-kpm-TimeEvOp.cpp"
+    # [LEGACY] "kpm-nonOrth-precision-test|kpm-nonOrth-precision-test.cpp"
     "inline_kpm-DOS-standalone|inline_kpm-DOS-standalone.cpp"
-    "device-postProcess|device-postProcess.cpp"
-    "inline_compute-Kubo-kpm-FFT-nonOrth|inline_compute-Kubo-kpm-FFT-nonOrth.cpp"
-    "inline_compute-kpm-TimeEvOpwithWF|inline_compute-kpm-TimeEvOpwithWF.cpp"
+    # [LEGACY] "device-postProcess|device-postProcess.cpp"
+    # [LEGACY] "inline_compute-Kubo-kpm-FFT-nonOrth|inline_compute-Kubo-kpm-FFT-nonOrth.cpp"
+    # [LEGACY] "inline_compute-kpm-TimeEvOpwithWF|inline_compute-kpm-TimeEvOpwithWF.cpp"
     "lsquant_inspect|lsquant_inspect.cpp"
     "lsquant|lsquant.cpp"
 )
